Wrap availability created events table behind a suspense boundary
